modified form

+++ Cheers-backend/Cheers/Models/Idea.cs
@@ -21,7 +21,7 @@ namespace Cheers.Models
         
         public Category Category { get; set; }
 
-        public long Target { get; set; }
+        public int Target { get; set; }
 
         public override string ToString()
         {
